+++ src/app/providers/app-config.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
-// import { HttpClient } from '@angular/common/http';
-import { Headers, Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
+// import { Headers, Http } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 
 import { environment } from '../../environments/environment';
@@ -13,12 +13,12 @@ import { getParameterByName } from '../utils/utils';
 export class AppConfigService {
   private appConfig;
 
-  constructor(private http: Http, public g: Globals) {
+  constructor(private http: HttpClient, public g: Globals) {
     this.appConfig = environment;
     console.log('chat21-web-widget environment: ', environment);
   }
 
-  loadAppConfig(): Observable<any> {
+  loadAppConfig() { // : Observable<any> {
     // START GET BASE URL and create absolute url of remoteConfigUrl //
     let urlConfigFile = this.appConfig.remoteConfigUrl;
     if (!this.appConfig.remoteConfigUrl.startsWith('http')) {
@@ -41,21 +41,22 @@ export class AppConfigService {
           }
       }
     }
-    // console.log("baseURL: ", this.g.baseLocation , this.appConfig.remoteConfigUrl );
+    // urlConfigFile = 'https://widget.inspired.edu.au/widget-config.json';
+    // console.log('baseURL: ', this.g.baseLocation , this.appConfig.remoteConfigUrl );
     // END GET BASE URL and create absolute url of remoteConfigUrl //
-    const headers = new Headers();
-    headers.append('Content-Type', 'application/json');
-    return this.http
-      .get(urlConfigFile, { headers })
-      .map((response) => response.json());
+    // const headers = new Headers();
+    // headers.append('Content-Type', 'application/json');
+    // return this.http
+    //   .get(urlConfigFile, { headers })
+    //   .map((response) => response.json());
 
-    // return this.http.get(urlConfigFile)
-    //   .toPromise()
-    //   .then(data => {
-    //     this.appConfig = data;
-    //   }).catch(err => {
-    //     // console.log('error loadAppConfig');
-    //   });
+    return this.http.get(urlConfigFile)
+      .toPromise()
+      .then(data => {
+        this.appConfig = data;
+      }).catch(err => {
+        console.log('error loadAppConfig', err);
+      });
   }
 
   getConfig() {